+++ tabbycat/api/serializers.py
@@ -14,20 +14,13 @@ from breakqual.models import BreakCategory
 from draw.models import Debate, DebateTeam
 from motions.models import Motion
 from participants.emoji import pick_unused_emoji
-from participants.models import Adjudicator, Institution, Speaker, SpeakerCategory, Team
+from participants.models import Adjudicator, Institution, Region, Speaker, SpeakerCategory, Team
+from privateurls.utils import populate_url_keys
 from results.mixins import TabroomSubmissionFieldsMixin
 from tournaments.models import Round, Tournament
 from venues.models import Venue, VenueCategory
 
-from .fields import (AdjudicatorFeedbackIdentityField, AnonymisingHyperlinkedTournamentRelatedField, MotionHyperlinkedIdentityField,
-    RoundHyperlinkedIdentityField, RoundHyperlinkedRelatedField, TournamentHyperlinkedIdentityField,
-    TournamentHyperlinkedRelatedField)
-
-
-def empty_without_null_validator(value):
-    if value == '':
-        raise serializers.ValidationError("Field cannot be an empty string. Use null.")
-    return value
+from . import fields
 
 
 class TournamentSerializer(serializers.ModelSerializer):
@@ -36,7 +29,7 @@ class TournamentSerializer(serializers.ModelSerializer):
         view_name='api-tournament-detail',
         lookup_field='slug', lookup_url_kwarg='tournament_slug')
 
-    current_rounds = TournamentHyperlinkedRelatedField(
+    current_rounds = fields.TournamentHyperlinkedRelatedField(
         view_name='api-round-detail', read_only=True, many=True,
         lookup_field='seq', lookup_url_kwarg='round_seq',
     )
@@ -91,21 +84,21 @@ class TournamentSerializer(serializers.ModelSerializer):
 
 class RoundSerializer(serializers.ModelSerializer):
     class RoundMotionsSerializer(serializers.ModelSerializer):
-        url = MotionHyperlinkedIdentityField(view_name='api-motion-detail')
+        url = fields.MotionHyperlinkedIdentityField(view_name='api-motion-detail')
 
         class Meta:
             model = Motion
             exclude = ('round',)
 
     class RoundLinksSerializer(serializers.Serializer):
-        pairing = TournamentHyperlinkedIdentityField(
+        pairing = fields.TournamentHyperlinkedIdentityField(
             view_name='api-pairing-list',
             lookup_field='seq', lookup_url_kwarg='round_seq')
 
-    url = TournamentHyperlinkedIdentityField(
+    url = fields.TournamentHyperlinkedIdentityField(
         view_name='api-round-detail',
         lookup_field='seq', lookup_url_kwarg='round_seq')
-    break_category = TournamentHyperlinkedRelatedField(
+    break_category = fields.TournamentHyperlinkedRelatedField(
         view_name='api-breakcategory-detail',
         queryset=BreakCategory.objects.all(),
         allow_null=True, required=False)
@@ -144,7 +137,7 @@ class RoundSerializer(serializers.ModelSerializer):
         return round
 
     def update(self, instance, validated_data):
-        motions_data = validated_data.pop('motions')
+        motions_data = validated_data.pop('motion_set')
         for motion in motions_data:
             try:
                 Motion.objects.update_or_create(round=instance, seq=motion.get('seq'), defaults={
@@ -159,7 +152,7 @@ class RoundSerializer(serializers.ModelSerializer):
 
 class MotionSerializer(serializers.ModelSerializer):
     class RoundsSerializer(serializers.ModelSerializer):
-        round = TournamentHyperlinkedRelatedField(view_name='api-round-detail',
+        round = fields.TournamentHyperlinkedRelatedField(view_name='api-round-detail',
             lookup_field='seq', lookup_url_kwarg='round_seq',
             queryset=Round.objects.all())
 
@@ -167,7 +160,7 @@ class MotionSerializer(serializers.ModelSerializer):
             model = Motion
             fields = ('round', 'seq')
 
-    url = MotionHyperlinkedIdentityField(view_name='api-motion-detail')
+    url = fields.MotionHyperlinkedIdentityField(view_name='api-motion-detail')
     rounds = RoundsSerializer(many=True, source='as_iterable')
 
     class Meta:
@@ -188,10 +181,10 @@ class MotionSerializer(serializers.ModelSerializer):
 class BreakCategorySerializer(serializers.ModelSerializer):
 
     class BreakCategoryLinksSerializer(serializers.Serializer):
-        eligibility = TournamentHyperlinkedIdentityField(
+        eligibility = fields.TournamentHyperlinkedIdentityField(
             view_name='api-breakcategory-eligibility')
 
-    url = TournamentHyperlinkedIdentityField(
+    url = fields.TournamentHyperlinkedIdentityField(
         view_name='api-breakcategory-detail')
 
     _links = BreakCategoryLinksSerializer(source='*', read_only=True)
@@ -205,10 +198,10 @@ class BreakCategorySerializer(serializers.ModelSerializer):
 class SpeakerCategorySerializer(serializers.ModelSerializer):
 
     class SpeakerCategoryLinksSerializer(serializers.Serializer):
-        eligibility = TournamentHyperlinkedIdentityField(
+        eligibility = fields.TournamentHyperlinkedIdentityField(
             view_name='api-speakercategory-eligibility', lookup_field='pk')
 
-    url = TournamentHyperlinkedIdentityField(
+    url = fields.TournamentHyperlinkedIdentityField(
         view_name='api-speakercategory-detail', lookup_field='pk')
     _links = SpeakerCategoryLinksSerializer(source='*', read_only=True)
 
@@ -235,7 +228,7 @@ class BaseEligibilitySerializer(serializers.ModelSerializer):
 
 class BreakEligibilitySerializer(BaseEligibilitySerializer):
 
-    team_set = TournamentHyperlinkedRelatedField(
+    team_set = fields.TournamentHyperlinkedRelatedField(
         many=True,
         queryset=Team.objects.all(),
         view_name='api-team-detail',
@@ -249,7 +242,7 @@ class BreakEligibilitySerializer(BaseEligibilitySerializer):
 
 class SpeakerEligibilitySerializer(BaseEligibilitySerializer):
 
-    speaker_set = TournamentHyperlinkedRelatedField(
+    speaker_set = fields.TournamentHyperlinkedRelatedField(
         many=True,
         queryset=Speaker.objects.all(),
         view_name='api-speaker-detail',
@@ -265,10 +258,11 @@ class SpeakerEligibilitySerializer(BaseEligibilitySerializer):
 class SpeakerSerializer(serializers.ModelSerializer):
 
     class LinksSerializer(serializers.Serializer):
-        checkin = TournamentHyperlinkedIdentityField(tournament_field='team__tournament', view_name='api-speaker-checkin')
+        checkin = fields.TournamentHyperlinkedIdentityField(tournament_field='team__tournament', view_name='api-speaker-checkin')
 
-    url = TournamentHyperlinkedIdentityField(tournament_field='team__tournament', view_name='api-speaker-detail')
-    categories = TournamentHyperlinkedRelatedField(
+    url = fields.TournamentHyperlinkedIdentityField(tournament_field='team__tournament', view_name='api-speaker-detail')
+    team = fields.TournamentHyperlinkedRelatedField(view_name='api-team-detail', queryset=Team.objects.all())
+    categories = fields.TournamentHyperlinkedRelatedField(
         many=True,
         view_name='api-speakercategory-detail',
         queryset=SpeakerCategory.objects.all(),
@@ -288,18 +282,27 @@ class SpeakerSerializer(serializers.ModelSerializer):
     class Meta:
         model = Speaker
         fields = ('url', 'id', 'name', 'gender', 'email', 'phone', 'anonymous', 'pronoun',
-                  'categories', 'url_key', '_links')
+                  'categories', 'url_key', '_links', 'team')
 
-    def validate_url_key(self, value):
-        return empty_without_null_validator(value)
+    def create(self, validated_data):
+        url_key = validated_data.pop('url_key', None)
+        if url_key is not None and len(url_key) != 0:  # Let an empty string be null for the uniqueness constraint
+            validated_data['url_key'] = url_key
+
+        speaker = super().create(validated_data)
+
+        if url_key is None:
+            populate_url_keys([speaker])
+
+        return speaker
 
 
 class AdjudicatorSerializer(serializers.ModelSerializer):
 
     class LinksSerializer(serializers.Serializer):
-        checkin = TournamentHyperlinkedIdentityField(view_name='api-adjudicator-checkin')
+        checkin = fields.TournamentHyperlinkedIdentityField(view_name='api-adjudicator-checkin')
 
-    url = TournamentHyperlinkedIdentityField(view_name='api-adjudicator-detail')
+    url = fields.TournamentHyperlinkedIdentityField(view_name='api-adjudicator-detail')
     institution = serializers.HyperlinkedRelatedField(
         allow_null=True,
         view_name='api-global-institution-detail',
@@ -311,12 +314,12 @@ class AdjudicatorSerializer(serializers.ModelSerializer):
         view_name='api-global-institution-detail',
         queryset=Institution.objects.all(),
     )
-    team_conflicts = TournamentHyperlinkedRelatedField(
+    team_conflicts = fields.TournamentHyperlinkedRelatedField(
         many=True,
         view_name='api-team-detail',
         queryset=Team.objects.all(),
     )
-    adjudicator_conflicts = TournamentHyperlinkedRelatedField(
+    adjudicator_conflicts = fields.TournamentHyperlinkedRelatedField(
         many=True,
         view_name='api-adjudicator-detail',
         queryset=Adjudicator.objects.all(),
@@ -351,11 +354,15 @@ class AdjudicatorSerializer(serializers.ModelSerializer):
                   'institution', 'base_score', 'trainee', 'independent', 'adj_core',
                   'institution_conflicts', 'team_conflicts', 'adjudicator_conflicts', 'url_key', '_links')
 
-    def validate_url_key(self, value):
-        return empty_without_null_validator(value)
-
     def create(self, validated_data):
+        url_key = validated_data.pop('url_key', None)
+        if url_key is not None and len(url_key) != 0:  # Let an empty string be null for the uniqueness constraint
+            validated_data['url_key'] = url_key
+
         adj = super().create(validated_data)
+
+        if url_key is None:  # If explicitly null (and not just an empty string)
+            populate_url_keys([adj])
 
         if adj.institution is not None:
             adj.adjudicatorinstitutionconflict_set.get_or_create(institution=adj.institution)
@@ -364,13 +371,18 @@ class AdjudicatorSerializer(serializers.ModelSerializer):
 
 
 class TeamSerializer(serializers.ModelSerializer):
-    url = TournamentHyperlinkedIdentityField(view_name='api-team-detail')
+    class TeamSpeakerSerializer(SpeakerSerializer):
+        class Meta(SpeakerSerializer.Meta):
+            fields = ('url', 'id', 'name', 'gender', 'email', 'phone', 'anonymous', 'pronoun',
+                      'categories', 'url_key', '_links')
+
+    url = fields.TournamentHyperlinkedIdentityField(view_name='api-team-detail')
     institution = serializers.HyperlinkedRelatedField(
         allow_null=True,
         view_name='api-global-institution-detail',
         queryset=Institution.objects.all(),
     )
-    break_categories = TournamentHyperlinkedRelatedField(
+    break_categories = fields.TournamentHyperlinkedRelatedField(
         many=True,
         view_name='api-breakcategory-detail',
         queryset=BreakCategory.objects.all(),
@@ -389,7 +401,7 @@ class TeamSerializer(serializers.ModelSerializer):
                   'institution_conflicts')
 
     def __init__(self, *args, **kwargs):
-        self.fields['speakers'] = SpeakerSerializer(*args, many=True, required=False, **kwargs)
+        self.fields['speakers'] = self.TeamSpeakerSerializer(*args, many=True, required=False, **kwargs)
 
         super().__init__(*args, **kwargs)
 
@@ -467,6 +479,7 @@ class TeamSerializer(serializers.ModelSerializer):
 
 class InstitutionSerializer(serializers.ModelSerializer):
     url = serializers.HyperlinkedIdentityField(view_name='api-global-institution-detail')
+    region = fields.CreatableSlugRelatedField(slug_field='name', queryset=Region.objects.all(), required=False)
 
     class Meta:
         model = Institution
@@ -474,12 +487,12 @@ class InstitutionSerializer(serializers.ModelSerializer):
 
 
 class PerTournamentInstitutionSerializer(InstitutionSerializer):
-    teams = TournamentHyperlinkedRelatedField(
+    teams = fields.TournamentHyperlinkedRelatedField(
         source='team_set',
         many=True,
         view_name='api-team-detail',
     )
-    adjudicators = TournamentHyperlinkedRelatedField(
+    adjudicators = fields.TournamentHyperlinkedRelatedField(
         source='adjudicator_set',
         many=True,
         view_name='api-adjudicator-detail',
@@ -496,15 +509,16 @@ class PerTournamentInstitutionSerializer(InstitutionSerializer):
 class VenueSerializer(serializers.ModelSerializer):
 
     class LinksSerializer(serializers.Serializer):
-        checkin = TournamentHyperlinkedIdentityField(view_name='api-venue-checkin')
+        checkin = fields.TournamentHyperlinkedIdentityField(view_name='api-venue-checkin')
 
-    url = TournamentHyperlinkedIdentityField(view_name='api-venue-detail')
-    categories = TournamentHyperlinkedRelatedField(
+    url = fields.TournamentHyperlinkedIdentityField(view_name='api-venue-detail')
+    categories = fields.TournamentHyperlinkedRelatedField(
         source='venuecategory_set', many=True,
         view_name='api-venuecategory-detail',
         queryset=VenueCategory.objects.all(),
     )
     display_name = serializers.ReadOnlyField()
+    external_url = serializers.URLField(source='url', required=False, allow_blank=True)
     _links = LinksSerializer(source='*', read_only=True)
 
     class Meta:
@@ -513,8 +527,8 @@ class VenueSerializer(serializers.ModelSerializer):
 
 
 class VenueCategorySerializer(serializers.ModelSerializer):
-    url = TournamentHyperlinkedIdentityField(view_name='api-venuecategory-detail')
-    venues = TournamentHyperlinkedRelatedField(
+    url = fields.TournamentHyperlinkedIdentityField(view_name='api-venuecategory-detail')
+    venues = fields.TournamentHyperlinkedRelatedField(
         many=True,
         view_name='api-venue-detail',
         queryset=Venue.objects.all(),
@@ -544,16 +558,16 @@ class BaseStandingsSerializer(serializers.Serializer):
 
 
 class TeamStandingsSerializer(BaseStandingsSerializer):
-    team = TournamentHyperlinkedRelatedField(view_name='api-team-detail', queryset=Team.objects.all())
+    team = fields.TournamentHyperlinkedRelatedField(view_name='api-team-detail', queryset=Team.objects.all())
 
 
 class SpeakerStandingsSerializer(BaseStandingsSerializer):
-    speaker = AnonymisingHyperlinkedTournamentRelatedField(view_name='api-speaker-detail', anonymous_source='anonymous')
+    speaker = fields.AnonymisingHyperlinkedTournamentRelatedField(view_name='api-speaker-detail', anonymous_source='anonymous')
 
 
 class RoundPairingSerializer(serializers.ModelSerializer):
     class DebateTeamSerializer(serializers.ModelSerializer):
-        team = TournamentHyperlinkedRelatedField(view_name='api-team-detail', queryset=Team.objects.all())
+        team = fields.TournamentHyperlinkedRelatedField(view_name='api-team-detail', queryset=Team.objects.all())
 
         class Meta:
             model = DebateTeam
@@ -561,9 +575,9 @@ class RoundPairingSerializer(serializers.ModelSerializer):
 
     class DebateAdjudicatorSerializer(serializers.Serializer):
         adjudicators = Adjudicator.objects.all()
-        chair = TournamentHyperlinkedRelatedField(view_name='api-adjudicator-detail', queryset=adjudicators)
-        panellists = TournamentHyperlinkedRelatedField(many=True, view_name='api-adjudicator-detail', queryset=adjudicators)
-        trainees = TournamentHyperlinkedRelatedField(many=True, view_name='api-adjudicator-detail', queryset=adjudicators)
+        chair = fields.TournamentHyperlinkedRelatedField(view_name='api-adjudicator-detail', queryset=adjudicators)
+        panellists = fields.TournamentHyperlinkedRelatedField(many=True, view_name='api-adjudicator-detail', queryset=adjudicators)
+        trainees = fields.TournamentHyperlinkedRelatedField(many=True, view_name='api-adjudicator-detail', queryset=adjudicators)
 
         def save(self, **kwargs):
             aa = kwargs['debate'].adjudicators
@@ -573,14 +587,22 @@ class RoundPairingSerializer(serializers.ModelSerializer):
             aa.save()
             return aa
 
-    url = RoundHyperlinkedIdentityField(view_name='api-pairing-detail')
-    venue = TournamentHyperlinkedRelatedField(view_name='api-venue-detail', queryset=Venue.objects.all())
+    url = fields.RoundHyperlinkedIdentityField(view_name='api-pairing-detail')
+    venue = fields.TournamentHyperlinkedRelatedField(view_name='api-venue-detail', queryset=Venue.objects.all())
     teams = DebateTeamSerializer(many=True, source='debateteam_set')
     adjudicators = DebateAdjudicatorSerializer()
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if not kwargs['context']['request'].user.is_staff:
+            self.fields.pop('bracket')
+            self.fields.pop('room_rank')
+            self.fields.pop('importance')
+            self.fields.pop('result_status')
+
     class Meta:
         model = Debate
-        fields = ('url', 'id', 'venue', 'teams', 'adjudicators', 'sides_confirmed')
+        exclude = ('round', 'flags')
 
     def create(self, validated_data):
         teams_data = validated_data.pop('debateteam_set')
@@ -616,17 +638,7 @@ class RoundPairingSerializer(serializers.ModelSerializer):
 
 
 class FeedbackQuestionSerializer(serializers.ModelSerializer):
-
-    class ChoicesField(serializers.Field):
-
-        def to_representation(self, value):
-            return value.split(AdjudicatorFeedbackQuestion.CHOICE_SEPARATOR)
-
-        def to_internal_value(self, data):
-            return AdjudicatorFeedbackQuestion.CHOICE_SEPARATOR.join(data)
-
-    url = TournamentHyperlinkedIdentityField(view_name='api-feedbackquestion-detail')
-    choices = ChoicesField()
+    url = fields.TournamentHyperlinkedIdentityField(view_name='api-feedbackquestion-detail')
 
     class Meta:
         model = AdjudicatorFeedbackQuestion
@@ -635,7 +647,7 @@ class FeedbackQuestionSerializer(serializers.ModelSerializer):
 
 class FeedbackSerializer(TabroomSubmissionFieldsMixin, serializers.ModelSerializer):
 
-    class SourceField(TournamentHyperlinkedRelatedField):
+    class SourceField(fields.TournamentHyperlinkedRelatedField):
         """Taken from REST_Framework: rest_framework.relations.HyperlinkedRelatedField
 
         This subclass adapts the framework in order to have a hyperlinked field which
@@ -702,12 +714,12 @@ class FeedbackSerializer(TabroomSubmissionFieldsMixin, serializers.ModelSerializ
             except self.model.DoesNotExist:
                 self.fail('does_not_exist')
 
-    class DebateHyperlinkedRelatedField(RoundHyperlinkedRelatedField):
+    class DebateHyperlinkedRelatedField(fields.RoundHyperlinkedRelatedField):
         def lookup_kwargs(self):
             return {self.tournament_field: self.context['tournament']}
 
     class FeedbackAnswerSerializer(serializers.Serializer):
-        question = TournamentHyperlinkedRelatedField(
+        question = fields.TournamentHyperlinkedRelatedField(
             view_name='api-feedbackquestion-detail',
             queryset=AdjudicatorFeedbackQuestion.objects.all(),
         )
@@ -719,8 +731,8 @@ class FeedbackSerializer(TabroomSubmissionFieldsMixin, serializers.ModelSerializ
             data['answer'] = model.ANSWER_TYPE(data['answer'])
             return super().validate(data)
 
-    url = AdjudicatorFeedbackIdentityField(view_name='api-feedback-detail')
-    adjudicator = TournamentHyperlinkedRelatedField(view_name='api-adjudicator-detail', queryset=Adjudicator.objects.all())
+    url = fields.AdjudicatorFeedbackIdentityField(view_name='api-feedback-detail')
+    adjudicator = fields.TournamentHyperlinkedRelatedField(view_name='api-adjudicator-detail', queryset=Adjudicator.objects.all())
     source = SourceField(source='*')
     debate = DebateHyperlinkedRelatedField(view_name='api-pairing-detail', queryset=Debate.objects.all())
     answers = FeedbackAnswerSerializer(many=True, source='get_answers', required=False)