+++ tabbycat/draw/models.py
@@ -1,5 +1,7 @@
 import logging
 
+from django.contrib.humanize.templatetags.humanize import ordinal
+from django.contrib.postgres.fields import ArrayField
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 from django.db import models
 from django.utils.translation import gettext, gettext_lazy as _
@@ -47,8 +49,7 @@ class Debate(models.Model):
     room_rank = models.IntegerField(default=0,
         verbose_name=_("room rank"))
 
-    # comma-separated list of strings
-    flags = models.CharField(max_length=100, blank=True)
+    flags = ArrayField(base_field=models.CharField(max_length=15), blank=True, choices=DRAW_FLAG_DESCRIPTIONS, default=list)
 
     importance = models.IntegerField(default=0, choices=[(i, i) for i in range(-2, 3)],
         verbose_name=_("importance"))
@@ -218,13 +219,6 @@ class Debate(models.Model):
                 self._confirmed_ballot = None
             return self._confirmed_ballot
 
-    def get_flags_display(self):
-        if not self.flags:
-            return []  # don't return [""]
-        else:
-            # If the verbose description can't be found, just show the raw flag
-            return [DRAW_FLAG_DESCRIPTIONS.get(f, f) for f in self.flags.split(",")]
-
     @property
     def history(self):
         try:
@@ -280,8 +274,7 @@ class DebateTeam(models.Model):
     side = models.CharField(max_length=3, choices=SIDE_CHOICES,
         verbose_name=_("side"))
 
-    # comma-separated list of strings
-    flags = models.CharField(max_length=100, blank=True)
+    flags = ArrayField(base_field=models.CharField(max_length=15), blank=True, choices=DRAW_FLAG_DESCRIPTIONS, default=list)
 
     class Meta:
         verbose_name = _("debate team")
@@ -303,23 +296,10 @@ class DebateTeam(models.Model):
                 self._opponent = None
             return self._opponent
 
-    def get_flags_display(self):
-        if not self.flags:
-            return []  # don't return [""]
-        else:
-            # If the verbose description can't be found, just show the raw flag
-            return [DRAW_FLAG_DESCRIPTIONS.get(f, f) for f in self.flags.split(",")]
-
     def get_result_display(self):
         if self.team.tournament.pref('teams_in_debate') == 'bp':
-            if self.points == 3:
-                return gettext("placed 1st")
-            elif self.points == 2:
-                return gettext("placed 2nd")
-            elif self.points == 1:
-                return gettext("placed 3rd")
-            elif self.points == 0:
-                return gettext("placed 4th")
+            if self.points is not None:
+                return gettext("placed %(place)s") % {'place': ordinal(4 - self.points)}
             else:
                 return gettext("result unknown")
         else: