+++ tabbycat/results/models.py
@@ -42,11 +42,16 @@ class Submission(models.Model):
     confirmed = models.BooleanField(default=False,
         verbose_name=_("confirmed"))
 
+    # relevant for private URL submissions
+    participant_submitter = models.ForeignKey('participants.Person', models.PROTECT,
+        blank=True, null=True, related_name="%(app_label)s_%(class)s_participant_submitted",
+        verbose_name=_("from participant"))
+
     # only relevant if submitter was in tab room
-    submitter = models.ForeignKey(settings.AUTH_USER_MODEL, models.CASCADE,
+    submitter = models.ForeignKey(settings.AUTH_USER_MODEL, models.PROTECT,
         blank=True, null=True, related_name="%(app_label)s_%(class)s_submitted",
         verbose_name=_("submitter"))
-    confirmer = models.ForeignKey(settings.AUTH_USER_MODEL, models.CASCADE,
+    confirmer = models.ForeignKey(settings.AUTH_USER_MODEL, models.PROTECT,
         blank=True, null=True, related_name="%(app_label)s_%(class)s_confirmed",
         verbose_name=_("confirmer"))
     confirm_timestamp = models.DateTimeField(blank=True, null=True,
@@ -229,7 +234,7 @@ class TeamScore(models.Model):
 
     points = models.PositiveSmallIntegerField(null=True, blank=True,
         verbose_name=_("points"))
-    win = models.NullBooleanField(null=True, blank=True,
+    win = models.BooleanField(null=True, blank=True,
         verbose_name=_("win"))
     margin = ScoreField(null=True, blank=True,
         verbose_name=_("margin"))