+++ tabbycat/templates/allocations/HoverPanelGroup.vue
@@ -1,7 +1,7 @@
 <template>
 
   <div class="btn-toolbar flex-nowrap">
-    <template v-if="group.length > 0">
+    <template v-if="groups.length > 0">
       <div v-for="group in groups" class="btn-group btn-group-sm mr-2">
         <button :class="['btn', item.css]" v-for="item in group">
           {{ item.title }}