+++ .github/CHANGELOG.rst
@@ -6,6 +6,31 @@ Change Log
 ----------------
 *Release date: ?*
 
+- Added URL field from the Room model to the Venues API endpoints as 'external URL'
+- The Institution API endpoints show institutions' regions as their name, and regions can be created
+- Adjudicator Feedback choices and draw flags are stored with PostgreSQL-specific arrays (`#1525 <https://github.com/TabbycatDebate/tabbycat/issues/1525>`_)
+- Private URL submissions now store the participant whose link was used rather than just their IP address (`#1586 <https://github.com/TabbycatDebate/tabbycat/issues/1586>`_)
+
+
+2.4.6
+-----
+*Release date: 19 September 2020*
+
+- Fixed issue where team names with an institution were longer than the maximum size (`#1564 <https://github.com/TabbycatDebate/tabbycat/issues/1564>`_)
+- Fixed issue that made saving motions impossible through the Round API detail endpoint
+- Fixed issue that made filtering by source team impossible for the Feedback API
+
+
+2.4.5
+-----
+*Release date: 19 July 2020*
+
+- Fixed the display of room URLs on private URL pages (thanks Viran for the report)
+- Preformed panels with a bracket range now sort more sensibly (thanks Zachary for the report)
+- Fixed manual sorting of preformed panels in general
+- Improved sorting of feedback table when sorting 'difference between base score and current score' (thanks Zachary for the report)
+- Fixed issue where the last saved counter was not updating on normal tables
+
 
 2.4.4
 -----